Asset Processor
Started the asset processor. The test bitmap(the only one...) shown in the misc color screen is now a binary format. I will provide documentation for these formats...but currently it is trivial. We store the width, height, and pixels. So, the layout:
|width|height|pixel0|pixel1|...|

I am preparing for 3d model import and would rather not have to mess around with the wavefront obj file format in the engine code. Additionally, I do not care much for text file formats; prefer to work with binary files... so the vertices, triangles, normals, uvs, and etc will all be written by the asset processor and the easily imported into the engine. The pre-processor will be uploaded in a day or two. I need to resolve some build script issues... pushd is giving me permissions denied stuff for some reason. I may just move the build file to the misc directory....but we will see.