Interpolated Directional Light
Fixed the the faceted surface. The normal interpolation was correct in the triangle shader... But the normals exported from blender were not interpolated. This resulted in all vertices for a given triangle having the same normal direction. So... another setting needed to be configures in Blender for proper interpolated normals to be exported.

I also added directional light color computations with a nice 1024x1024 brick texture for the sphere. The texture size is probably overkill but why not??? =D

Texture source: https://textures.pixel-furnace.com/