Normal Mapping!
We added normal mapping. Normal mapping happens in tangent  space, so we needed to compute tangents and bitangents. We only need to store tangents. Bitangents cant be reconstruct with the normal and tangent, due to orthogonality. Updated mesh data format to include tangents.

Yet, some of the calculations may be wrong, as specular lighting is broken with normals. Yet, there are updated screens that show the normal mapping with diffuse lighting.