Normal Mapping Complete?
We believe to have proper normal mapping now. The results are satisfying - screens in misc directory updated.

There were a number of initial problems with normal mapping. First, I thought the normal map was 32 bit depth...so importing a png and exporting a bmp for our asset processor was exporting 32 bit depth pixels. Yet, the png normal was 24 bit. Corrected that in the asset processor. Additionally I changed the mesh file format. We no longer have trianges. Instead we construct vertex attributes and the tangent vectors. Then we dump it all. This increased the file size, and I am not sure if an array of structs is the right way to go. Yet, it will remain for now and we can reevaluate when we optimize. File format docs have been updated.

Added the concept of a material and pulled the vertex shader out into a function so I can render more that one mesh with ease - preparation for shadow mapping.

Still unsure how I want to handle clipping. So, do not show a mesh your backside or it will give you a swift kick in the.....